backend_api/business_card_management_system/requirements.txt
asgiref==3.8.1
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
numpy==1.26.4
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
pytesseract==0.3.10
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
